Refactor Scaffolded Classes & EFCF InitialMigration
