+++ NewsFeeds/Models/ApplicationDbContext.cs
@@ -1,0 +1,17 @@
+﻿using Microsoft.AspNet.Identity.EntityFramework;
+
+namespace NewsFeeds.Models
+{
+    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
+    {
+        public ApplicationDbContext()
+            : base("DefaultConnection", throwIfV1Schema: false)
+        {
+        }
+
+        public static ApplicationDbContext Create()
+        {
+            return new ApplicationDbContext();
+        }
+    }
+}
