+++ NewsFeeds/Views/Account/ExternalLoginConfirmation.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.ExternalLoginConfirmationViewModel
+﻿@model NewsFeeds.ViewModels.ExternalLoginConfirmationViewModel
 @{
     ViewBag.Title = "Register";
 }
