+++ NewsFeeds/Views/Account/ForgotPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.ForgotPasswordViewModel
+﻿@model NewsFeeds.ViewModels.ForgotPasswordViewModel
 @{
     ViewBag.Title = "Forgot your password?";
 }
