+++ NewsFeeds/Views/Account/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@using NewsFeeds.Models
+﻿@using NewsFeeds.ViewModels
 @model LoginViewModel
 @{
     ViewBag.Title = "Log in";
