+++ NewsFeeds/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.RegisterViewModel
+﻿@model NewsFeeds.ViewModels.RegisterViewModel
 @{
     ViewBag.Title = "Register";
 }
