+++ NewsFeeds/Views/Account/ResetPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.ResetPasswordViewModel
+﻿@model NewsFeeds.ViewModels.ResetPasswordViewModel
 @{
     ViewBag.Title = "Reset password";
 }
