+++ NewsFeeds/Views/Account/SendCode.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.SendCodeViewModel
+﻿@model NewsFeeds.ViewModels.SendCodeViewModel
 @{
     ViewBag.Title = "Send";
 }
