+++ NewsFeeds/Views/Account/VerifyCode.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.VerifyCodeViewModel
+﻿@model NewsFeeds.ViewModels.VerifyCodeViewModel
 @{
     ViewBag.Title = "Verify";
 }
