+++ NewsFeeds/Views/Account/_ExternalLoginsListPartial.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.ExternalLoginListViewModel
+﻿@model NewsFeeds.ViewModels.ExternalLoginListViewModel
 @using Microsoft.Owin.Security
 
 <h4>Use another service to log in.</h4>
