+++ NewsFeeds/Views/Manage/AddPhoneNumber.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.AddPhoneNumberViewModel
+﻿@model NewsFeeds.ViewModels.AddPhoneNumberViewModel
 @{
     ViewBag.Title = "Phone Number";
 }
