+++ NewsFeeds/Views/Manage/ChangePassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.ChangePasswordViewModel
+﻿@model NewsFeeds.ViewModels.ChangePasswordViewModel
 @{
     ViewBag.Title = "Change Password";
 }
