+++ NewsFeeds/Views/Manage/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.IndexViewModel
+﻿@model NewsFeeds.ViewModels.IndexViewModel
 @{
     ViewBag.Title = "Manage";
 }
