+++ NewsFeeds/Views/Manage/ManageLogins.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.ManageLoginsViewModel
+﻿@model NewsFeeds.ViewModels.ManageLoginsViewModel
 @using Microsoft.Owin.Security
 @{
     ViewBag.Title = "Manage your external logins";
