+++ NewsFeeds/Views/Manage/SetPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.SetPasswordViewModel
+﻿@model NewsFeeds.ViewModels.SetPasswordViewModel
 @{
     ViewBag.Title = "Create Password";
 }
