+++ NewsFeeds/Views/Manage/VerifyPhoneNumber.cshtml
@@ -1,4 +1,4 @@
-﻿@model NewsFeeds.Models.VerifyPhoneNumberViewModel
+﻿@model NewsFeeds.ViewModels.VerifyPhoneNumberViewModel
 @{
     ViewBag.Title = "Verify Phone Number";
 }
